Support for flash.

+++ app/src/commonMain/kotlin/com/eightsines/bpe/util/ElapsedTimeProvider.kt
@@ -1,5 +1,5 @@
 package com.eightsines.bpe.util
 
 interface ElapsedTimeProvider {
-    fun getElapsedTimeMs(): Int
+    fun getElapsedTimeMs(): Long
 }
